requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
Cython==0.29.7
entrypoints==0.3
flake8==3.7.7
future==0.17.1
idna==2.8
isort==4.3.18
lazy-object-proxy==1.3.1
lxml==4.3.3
mccabe==0.6.1
psycopg2-binary==2.8.3
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pylint==2.3.1
PySocks==1.7.0
python-telegram-bot==11.1.0
rauth==0.7.3
requests==2.21.0
six==1.12.0
typed-ast==1.3.5
urllib3==1.24.3
vedis==0.7.1
wrapt==1.11.1
xmltodict==0.12.0
